refactor: move poi-list-elements to tds-widget
