--- packages/poi-list-elements/package.json
@@ -1,64 +1,0 @@
-{
-  "name": "@titicaca/poi-list-elements",
-  "version": "13.4.0",
-  "description": "Triple POI UI Elements",
-  "license": "MIT",
-  "homepage": "https://github.com/titicacadev/triple-frontend/tree/main/packages/poi-list-elements",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/titicacadev/triple-frontend.git"
-  },
-  "bugs": {
-    "url": "https://github.com/titicacadev/triple-frontend/issues"
-  },
-  "main": "src/index.ts",
-  "types": "src/index.ts",
-  "publishConfig": {
-    "main": "lib/index.js",
-    "types": "lib/index.d.ts"
-  },
-  "files": [
-    "lib"
-  ],
-  "scripts": {
-    "build": "pnpm run build:cjs && pnpm run build:types",
-    "build:cjs": "swc src -d lib --config-file ../../.swcrc",
-    "build:types": "tsc -p tsconfig.build.json",
-    "lint:es": "eslint src",
-    "lint:es:fix": "eslint src --fix",
-    "lint:style": "stylelint 'src/**/*.{js,ts,tsx}'",
-    "lint:style:fix": "stylelint 'src/**/*.{js,ts,tsx}' --fix",
-    "lint:etc": "prettier src --check",
-    "lint:etc:fix": "prettier src --write"
-  },
-  "lint-staged": {
-    "*": [
-      "prettier --check"
-    ],
-    "*.{js,ts,tsx}": [
-      "eslint"
-    ]
-  },
-  "dependencies": {
-    "@titicaca/carousel": "workspace:*",
-    "@titicaca/core-elements": "workspace:*",
-    "@titicaca/resource-list-element": "workspace:*",
-    "@titicaca/scrap-button": "workspace:*",
-    "@titicaca/type-definitions": "workspace:*",
-    "@titicaca/view-utilities": "workspace:*"
-  },
-  "devDependencies": {
-    "@titicaca/i18n": "workspace:*",
-    "@titicaca/next-i18next": "13.8.5",
-    "@titicaca/react-contexts": "workspace:*",
-    "react": "^18.2.0",
-    "styled-components": "^5.3.11"
-  },
-  "peerDependencies": {
-    "@titicaca/i18n": "*",
-    "@titicaca/next-i18next": "*",
-    "@titicaca/react-contexts": "*",
-    "react": "^18.0",
-    "styled-components": "^5.3.9"
-  }
-}
