--- packages/poi-list-elements/src/styled.d.ts
@@ -1,1 +1,0 @@
-import {} from 'styled-components/cssprop'
