--- packages/poi-list-elements/tsconfig.build.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "declaration": true,
-    "emitDeclarationOnly": true,
-    "noEmitOnError": true,
-    "outDir": "./lib"
-  },
-  "include": ["src/**/*"],
-  "exclude": [
-    "node_modules",
-    "lib",
-    "src/**/*.test.*",
-    "src/**/*.spec.*",
-    "src/**/*.stories.*"
-  ]
-}
