--- packages/poi-list-elements/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "exclude": ["node_modules", "lib"]
-}
