+++ packages/tds-widget/src/poi-list-elements/compact-poi-list-element.tsx
