+++ packages/tds-widget/src/poi-list-elements/extended-poi-list-element.tsx
