+++ packages/tds-widget/src/poi-list-elements/get-type-names.ts
