+++ packages/tds-widget/src/poi-list-elements/mocks/hotels.sample.json
