+++ packages/tds-widget/src/poi-list-elements/mocks/pois.sample.json
